Move injected client to package:dwds
