+++ _test/pubspec.yaml
@@ -1,0 +1,23 @@
+name: _test
+# Every time this changes you need to run `pub run build_runner build`.
+version: 1.0.0
+description: >-
+  A fake package used for testing
+
+environment:
+  sdk: ">=2.3.2-dev.0.1 <3.0.0"
+
+dependencies:
+  intl: ^0.15.8
+  path: ^1.6.1
+  webdev: ^2.0.0
+
+dev_dependencies:
+  build_runner: ^1.0.0
+  build_web_compilers: '>=1.0.0 <3.0.0'
+
+dependency_overrides:
+  webdev:
+    path: ../webdev
+  dwds:
+    path: ../dwds
