+++ dwds/lib/src/injected/configuration.dart
