+++ webdev/lib/src/command/configuration.dart
@@ -3,10 +3,10 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:args/args.dart';
+import 'package:dwds/src/injected/configuration.dart'; // ignore: implementation_imports
 import 'package:logging/logging.dart';
 
 import '../logging.dart';
-import '../serve/injected/configuration.dart';
 
 const autoOption = 'auto';
 const chromeDebugPortFlag = 'chrome-debug-port';
