+++ webdev/pubspec.yaml
@@ -38,7 +38,6 @@ dependencies:
 dev_dependencies:
   build: ^1.0.0
   build_runner: ^1.0.0
-  build_web_compilers: '>=1.0.0 <3.0.0'
   build_verify: ^1.0.0
   build_version: ^2.0.0
   graphs: ^0.2.0
